refactor: consolidate articulation ordering helpers
